Add files via upload
16 Dec 2021
PHP 7.4.27 Released!
The PHP development team announces the immediate availability of PHP 7.4.27. This is a bug fix release.

All PHP 7.4 users are encouraged to upgrade to this version.

For source downloads of PHP 7.4.27 please visit our downloads page, Windows source and binaries can be found on windows.php.net/download/. The list of changes is recorded in the ChangeLog.